Added more features to both, the client and server side of the program
Server side: added static final String fields that identify a 3
character code that the server uses to find out what kind of a message
is being received from the client and act accordingly.
Client side: Changed it so that every message sent to the server would
include the 3 character code (using the server.Server's static final
fields) in order to set the client's username, send a disconnect signal,
or send a normal message.